MO-11617: Improve determinism by making the tracker wait until the mapper has finished.
The tracker passes keyframes to the mapping thread, which compares and tidies up landmarks, deleting/merging ones that represent the same point. If the tracker is allowed to continue while this is happening, it will produce different results depending on whether it performs certain actions before or after landmarks have been merged.

Making it wait in this way somewhat defeats the point of the separate threads; this should probably be revisited in future to insert the waits in more specific locations.