Fixed Windows compiler options when building a shared library. Now uses /MD instead of /MT.
OpenMP is disabled if USE_OPENMP is defined and set to a false value.
Prior to this, OpenMP could not be disabled.

Fixed a defect in finding OpenCV 4 on Windows if OpenCV 3 is not found.

Only set CMAKE_CXX_STANDARD to use c++11 if the variable is not already set.