Reset static next_id_ value for frame, keyframe and landmark IDs.
Keeping the keyframe ids consistent is necessary for deterministic behaviour when the solve is run multiple times in the same session, because the keyframe with ID 0 is treated as "fixed" in various places (look for "id_ == 0").

The frame and landmark IDs shouldn't affect the solve results and are reset for consistency/debugging.